Update to use Cyclone's renamed ddsi_sertype
The Cyclone changes are binary compatible (if with a little bit of
overhead) but not quite source compatible.  Source compatibility is
problematic in C because the signatures of some functions change as a
consequence of (finally) moving the topic name out of the type support.
Providing compatibility has been looked into and is feasible, but would
not really bring any benefit because of the overhead.

Signed-off-by: Erik Boasson <eb@ilities.com>