Report history depth for keep-all as 0
For a keep-all history the depth setting is irrelevant, but the XML-RPC
used between the ros2cli daemon and ros2cli client doesn't deal
gracefully with integers >= 2**31.  Returning 0 instead of the value
returned by the Cyclone get_qos operation works around that issue
without affecting the meaning of the result.

Signed-off-by: Erik Boasson <eb@ilities.com>